Few fixes
Fixed the background image issue with it being offset on smaller screens once you get down to phone sizes under a certain pixel range. The image was showing the far left of the image. I decided to have the image stretch instead. It's a large image, and the stretching does not seem to effect it to much, because the breakpoints cause the image to reset at the points I set. So the stretching is not an issue, and looks much more appealing on all screen sizes.